More work on KNX backend. Siwtch and light implemented
KNCtrl is spawning 2 processes, 1 for sending commands to eibnetmux and
the other to listen and monitor the bus. This has to be like that
because the monitor function is blocking, and we cannot send command
when the connection is in monitor mode.
Another specific feature is when eibnetmux daemon exits, only the
monitor process knows that the connection fails. It then sends a message
to KNXCtrl "disconnect" so that it can terminate() the command process
and restart properly. The command process has no way to detect a
connection drop beacuse of the design of eibnetmux library.